+++ benchmarks/handmade-1/handmade1.py
@@ -1,19 +1,20 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 # @Project: PyFuzzSDE
-# @File:    pyfuzzsde/benchmarks/benchmark_01.py
-# @Time:    2021/04/21 03:00:00
-# @Version: 0.0.3
+# @File:    pyfuzzsde/benchmarks/handmade-1/handmade1.py
+# @Time:    2021/05/18 16:00:00
+# @Version: 0.0.4
 # @Author:  LightningRS
 # @Email:   me@ldby.site
-# @Desc:    PyFuzzSDE Benchmark 01
+# @Desc:    PyFuzzSDE Handmade Benchmark 01
 # @Changelog:
+#    2021/05/18: 调整文件结构
 #    2021/04/21: 微调用例用于验证动态分析结果
 #    2021/04/21: Create benchmark
 
 FLAGS_COUNT = 11
 
-def benchmark_01(s: str):
+def handmade_1(s: str):
     suffix = s[-7:]
     print("suffix = {}".format(suffix))
     if s == 'bad':